web/requirements.txt
Django==3.2.4
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.4
psycopg2==2.8.6
djangorestframework==3.12.4
django-elasticsearch-dsl==7.2.0
django-rest-elasticsearch==0.4.2
sshtunnel==0.4.0
sqlalchemy==1.4.26
pyvis==0.1.9

